+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Dev: Improve Wikipedia tests. (#286)
 - Dev: Improve Imgur tests. (#289)
 - Dev: Improve migration tests. (#290)
+- Dev: Improve Twitter tests. (#293)
 
 ## 1.2.3